+++ FE/src/App.tsx
@@ -32,7 +32,6 @@ import CartContext from "./contexts/CartContext";
 import OrderContext from "./contexts/OrderContext";
 import ThankYou from "./pages/client/Thankyou";
 import Contact from "./pages/client/contact";
-import Test from "./pages/client/test";
 
 function App() {
   return (
@@ -53,15 +52,12 @@ function App() {
           }
         >
           <Route index element={<Home />} />
-
           <Route path="/detail/:id" element={<Detail />} />
           <Route path="/products" element={<Products />} />
           <Route path="/checkout" element={<Checkout />} />
           <Route path="/cart" element={<Cart />} />
           <Route path="/contact" element={<Contact />} />
           <Route path="/order-history" element={<OrderHistory />} />
-          <Route path="/test" element={<Test />} />
-
           <Route path="/login" element={<AuthForm isLogin />} />
           <Route path="/register" element={<AuthForm />} />
           <Route path="/thankyou" element={<ThankYou />} />
