+++ FE/src/pages/admin/order/Order.tsx
@@ -44,11 +44,12 @@ const AdminOrder = () => {
 
   const getStatusOptions = (currentStatus: string) => {
     const statusFlow = [
-      "chờ xử lý",
-      "đã xác nhận",
-      "đang vận chuyển",
-      "hoàn thành",
-      "đã hủy",
+      "Chờ xử lý",
+      "Đã xác nhận",
+      "Đang vận chuyển",
+      "Đã giao hàng",
+      "Hoàn thành",
+      "Đã hủy",
     ];
 
     const currentIndex = statusFlow.indexOf(currentStatus);
